육친 변환시 한자 金이 인식되지 않는 오류 수정(다른 코드의 金이 섞여있었음)

+++ src/furtune/Api_Lunar_XmlParser.java
@@ -12,7 +12,7 @@ import java.io.StringReader;
 public class Api_Lunar_XmlParser {
 	
 	
-	public void parseXmlString(String xmlStr, int inputHour) throws Exception {
+	public void parseXmlString(String xmlStr, double inputHour) throws Exception {
         DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
         DocumentBuilder builder = factory.newDocumentBuilder();
         InputSource is = new InputSource(new StringReader(xmlStr));
